backend/requirements.txt
kafka-python==2.0.1
pymongo==3.10.1
python-dateutil==2.8.1
six==1.15.0
pymongo==3.10.1